requirements.txt
aniso8601==9.0.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.7

pytest-flask==1.2.0
pytz==2021.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
tomli==1.2.3
urllib3==1.26.8
Werkzeug==2.0.3
